+++ src/frontend/apps/desk/src/core/MainLayout.tsx
@@ -5,13 +5,17 @@ import { Footer } from '@/features/footer/Footer';
 import { HEADER_HEIGHT, Header } from '@/features/header';
 import { Menu } from '@/features/menu';
 
+import { useConfigStore } from './config';
+
 export function MainLayout({ children }: PropsWithChildren) {
+  const { config } = useConfigStore();
+
   return (
     <Box>
       <Box $height="100vh">
         <Header />
         <Box $css="flex: 1;" $direction="row">
-          {process.env.NEXT_PUBLIC_FEATURE_TEAM === 'true' && <Menu />}
+          {config?.FEATURES.TEAMS && <Menu />}
           <Box
             as="main"
             $height={`calc(100vh - ${HEADER_HEIGHT})`}